requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.17
albumentations==1.4.18
altair==5.4.1
annotated-types==0.7.0
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
efficientnet_pytorch==0.7.1
eval_type_backport==0.2.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fsspec==2024.9.0
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.2
h5py==3.12.1
idna==3.10
iglovikov-helper-functions==0.0.53
imagecorruptions==1.1.2
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jpeg4py==0.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.9.3
networkx==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==4.25.5
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
retinaface-pytorch==0.0.7
rich==13.9.2
rpds-py==0.20.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.39.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
waitress==3.0.0
watchdog==5.0.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
timm